feat: move into gem and module structure, fixes #4, fixes #8, fixes #9
