+++ Gemfile
@@ -1,0 +1,8 @@
+source "https://rubygems.org"
+
+# Specify your gem's dependencies in relaton_nist.gemspec
+gemspec
+
+gem "rubocop-performance"
+gem "rubocop-rake"
+gem "rubocop-rspec"
