+++ lib/atmospheric.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require_relative "atmospheric/version"
+require_relative "atmospheric/isa"
+
+module Atmospheric
+  class Error < StandardError; end
+  # Your code goes here...
+end
