+++ lib/atmospheric/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module Atmospheric
+  VERSION = "0.1.0"
+end
