+++ sig/atmospheric.rbs
@@ -1,0 +1,4 @@
+module Atmospheric
+  VERSION: String
+  # See the writing guide of rbs: https://github.com/ruby/rbs#guides
+end
