+++ spec/atmospheric_spec.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+RSpec.describe Atmospheric do
+  it "has a version number" do
+    expect(Atmospheric::VERSION).not_to be nil
+  end
+
+end
